refactor(new post): put new post form in its own file
